web_api_service/requirements.txt
absl-py==0.11.0
amqp==5.0.2
asgiref==3.2.10
astunparse==1.6.3
billiard==3.6.3.0
cachetools==4.2.0
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dateparser==1.0.0
DAWG-Python==0.7.2
dicttoxml==1.7.4
Django==3.1.2
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-imagekit==4.0.2
djangorestframework==3.12.1
docopt==0.6.2
drf-yasg==1.20.0
gast==0.3.3
google-api-core==1.24.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-vision==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
inflection==0.5.1
intervaltree==3.1.0
ipymarkup==0.9.0
itypes==1.2.0
Jinja2==2.11.2
joblib==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.0.2
libcst==0.3.15
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mypy-extensions==0.4.3
mysqlclient==2.0.1
natasha==1.4.0
navec==0.10.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pdf2image==1.14.0
pika==1.1.0
pilkit==2.0
Pillow==8.0.1
prompt-toolkit==3.0.8
proto-plus==1.13.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyMySQL==0.10.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
PyYAML==5.3.1
razdel==0.5.0
redis==3.5.3
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
slovnet==0.5.0
sortedcontainers==2.3.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.8.0
tensorflow-text==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yargy==0.15.0
